requirements.txt
fastapi
sqlmodel
uvicorn==0.30.1
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
psycopg2-binary
email-validator
